Option to See Unlimited Results
Previously, search results were limited to the top 200 (to save on load times). Now there is an option in settings to show every result.

The Default behavior is to limit results.

Also added a Try{}Catch{} to the sorting methods to catch (and ignore) errors sorting the Quantity column, (which will be dealt with at a later update).